requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
openai>=1.100.0
pydantic>=2.10.0
python-multipart>=0.0.12
python-dotenv>=0.21.0
pytest>=8.0.0
